requirements.txt
# Core dependencies
torch==2.0.1
torchvision==0.15.2
scikit-learn>=1.3.0,<=1.4.0
pandas>=2.0.0,<=2.1.4
numpy>=1.24.0,<=1.26.2
scipy>=1.10.0,<=1.11.4
attrs>=23.0.0
flatten_dict>=0.4.2
tensorboardX>=2.6.0
gcsfs>=2023.6.0
google-auth>=2.23.0
fsspec>=2023.6.0

# Testing
pytest>=7.4.0

# Lint and code style
flake8>=6.0.0
flake8-blind-except>=0.2.1
flake8-debugger>=4.1.2
flake8-print>=5.0.0
flake8-tuple>=0.4.1
mypy>=1.5.0